requirements.txt
pydantic-ai
pydantic
rich
python-dotenv
fastapi
uvicorn
jinja2
